+++ src/quantium/units/registry.py
@@ -41,7 +41,7 @@ from quantium.core.dimensions import (
     dim_mul,
     dim_pow,
 )
-from quantium.core.quantity import Unit
+from quantium.core.unit import Unit
 from quantium.units.parser import extract_unit_expr
 
 
@@ -218,7 +218,7 @@ class UnitsRegistry:
         Raises `ValueError` if unknown.
         """
         # if it is a composed expression
-        if any(op in symbol for op in ('*', '/')):
+        if any(op in symbol for op in ('*', '/', '^')):
             return extract_unit_expr(symbol, self)
         
         sym = normalize_symbol(symbol)