+++ tests/core/quantity/test_quantity_basics.py
@@ -3,7 +3,7 @@ import pytest
 from quantium.core.dimensions import LENGTH, TEMPERATURE,TIME, DIM_0
 from quantium.core.quantity import Quantity, Unit
 from quantium.units.registry import DEFAULT_REGISTRY as dreg
-from quantium import u
+from quantium.units import u
 # -------------------------------
 # Quantity: basics & conversion
 # -------------------------------